backend/requirements.txt
fastapi==0.115.7
fastapi-cli==0.0.7
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6emoviepy==2.1.2
moviepy==2.1.2langchain==0.3.15
langchain-cli==0.0.35
langchain-community==0.3.15
langchain-core==0.3.31
langchain-openai==0.3.1
langchain-text-splitters==0.3.5
chromadb==0.6.3
